Organize download classes under one folder
