+++ firefox-ios/Client/Frontend/Browser/DownloadHelper/DownloadToast.swift
